Week 04/backend/requirements.txt
Flask==2.0.1
Flask-JWT-Extended==4.3.1
Flask-CORS==3.0.10
python-dotenv==0.19.0
firebase-admin==5.0.3
PyJWT==2.8.0
cryptography==42.0.2
bcrypt==4.3.0
Werkzeug==3.0.1
gunicorn==21.2.0
python-dateutil==2.8.2
requests==2.32.3
urllib3==2.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pillow==8.3.2
qrcode==7.3.1
opencv-python==4.5.3.56
numpy==1.21.2
python-magic==0.4.24
pytesseract==0.3.8
scikit-image==0.18.3